v4.0 添加借书，还书模块，以及分页查询功能

+++ src/main/java/com/zihuv/managebooks/entity/Book.java
@@ -15,6 +15,6 @@ public class Book {
     private String bookAuthor;
     private String bookPublish;
     private Integer bookCategory;
-    private Integer bookPrize;
+    private Double bookPrize;
     private String bookIntroduction;
 }
